# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/user/{userId}/get_catalogs/catalog/updateName":{"put":{"tags":["user-controller"],"operationId":"updateCatalogName","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CatalogDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CatalogDto"}}}}}}},"/user/{userId}/get_catalogs/catalog/remove_admin":{"put":{"tags":["user-controller"],"operationId":"removeAdmin","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"catalogId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"removableAdminLogin","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/user/{userId}/get_catalogs/catalog/groups/enroll_group":{"put":{"tags":["user-controller"],"operationId":"enrollGroup","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"groupId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/user/{userId}/get_catalogs/catalog/group":{"put":{"tags":["user-controller"],"operationId":"getGroupEvents","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"groupId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventDto"}}}}}}}},"/user/{userId}/get_catalogs/catalog/group/edit_event":{"put":{"tags":["user-controller"],"operationId":"editEvent","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventDto"}}}}}}},"/user/{userId}/get_catalogs/catalog/add_admin":{"put":{"tags":["user-controller"],"operationId":"addAdmin","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"catalogId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"newAdminLogin","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/user/{userId}/enroll_catalog":{"put":{"tags":["user-controller"],"operationId":"enrollCatalog","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"catalogCode","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/user/{userId}/get_catalogs/create":{"post":{"tags":["user-controller"],"operationId":"createUserRootCatalog","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CatalogDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CatalogDto"}}}}}}},"/user/{userId}/get_catalogs/catalog/group/add_event":{"post":{"tags":["user-controller"],"operationId":"addEvent","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"groupId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventDto"}}}}}}},"/user/{userId}/get_catalogs/catalog/create_group":{"post":{"tags":["user-controller"],"operationId":"createGroup","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"catalogId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupDto"}}}}}}},"/user/{userId}/get_catalogs/catalog/create":{"post":{"tags":["user-controller"],"operationId":"createChildCatalog","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CatalogDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CatalogDto"}}}}}}},"/register":{"post":{"tags":["register-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/user/{userId}/schedule":{"get":{"tags":["user-controller"],"operationId":"getUserSchedule","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"startDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventDto"}}}}}}}},"/user/{userId}/get_catalogs":{"get":{"tags":["user-controller"],"operationId":"getUserRootCatalogs","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CatalogDto"}}}}}}}},"/user/{userId}/get_catalogs/catalog":{"get":{"tags":["user-controller"],"operationId":"getChildCatalogs","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"catalogId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CatalogDto"}}}}}}}},"/user/{userId}/get_catalogs/catalog/groups":{"get":{"tags":["user-controller"],"operationId":"getGroupsInCatalog","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"catalogId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupDto"}}}}}}}},"/user/{userId}/get_catalogs/catalog/get_admins":{"get":{"tags":["user-controller"],"operationId":"getAdmins","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"catalogId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/login":{"get":{"tags":["login-controller"],"operationId":"login","parameters":[{"name":"userDto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/UserDto"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/user/{userId}/get_catalogs/catalog/remove":{"delete":{"tags":["user-controller"],"operationId":"removeCatalog","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"catalogId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/user/{userId}/get_catalogs/catalog/group/remove_event":{"delete":{"tags":["user-controller"],"operationId":"removeEvent","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"eventId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/user/{userId}/get_catalogs/catalog/group/remove":{"delete":{"tags":["user-controller"],"operationId":"removeGroup","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"groupId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"CatalogDto":{"required":["name"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"ownerID":{"type":"string","format":"uuid"},"parentCatalogID":{"type":"string","format":"uuid"},"parentCatalogName":{"type":"string"},"code":{"type":"string"},"groups":{"type":"array","items":{"$ref":"#/components/schemas/GroupDto"}}}},"EventDto":{"required":["endTime","name","startTime"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"}}},"GroupDto":{"required":["name"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"events":{"type":"array","items":{"$ref":"#/components/schemas/EventDto"}}}},"UserCatalogDto":{"required":["catalogID","userID"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userID":{"type":"string","format":"uuid"},"catalogID":{"type":"string","format":"uuid"},"amin":{"type":"boolean"}}},"UserDto":{"required":["login","password"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"login":{"type":"string"},"password":{"type":"string"},"userCatalogList":{"type":"array","items":{"$ref":"#/components/schemas/UserCatalogDto"}},"groups":{"type":"array","items":{"$ref":"#/components/schemas/GroupDto"}}}}}}}